Rename pkg/context to k0scontext to avoid named imports
Since stdlib's `"context"` is imported almost everywhere, having an
internal package called `context` forces almost always importing it with
a label like `k0sctx`.

Here it is renamed to `k0scontext` to avoid having to do so.
