+++ pkg/k0scontext/context.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package context
+package k0scontext
 
 import (
 	"context"
